test(cli): split semantic cancellation tests
